+++ src/main/java/entity/CongThucMonAnEntity.java
@@ -1,7 +1,5 @@
 package entity;
 
-import java.math.BigDecimal;
-
 /**
  * Entity class: CongThucMonAnEntity
  */
@@ -9,7 +7,7 @@ public record CongThucMonAnEntity(
         int id,
         int monAnId,
         int nguyenLieuId,
-        BigDecimal khoiLuong,
+        Double khoiLuong,
         // Display fields (not in the database table but joined)
         String tenMon,
         String tenNguyenLieu,
@@ -18,7 +16,7 @@ public record CongThucMonAnEntity(
     /**
      * Constructor with only the database fields (no display fields)
      */
-    public CongThucMonAnEntity(int id, int monAnId, int nguyenLieuId, BigDecimal khoiLuong) {
+    public CongThucMonAnEntity(int id, int monAnId, int nguyenLieuId, Double khoiLuong) {
         this(id, monAnId, nguyenLieuId, khoiLuong, null, null, null);
     }
 }
