+++ src/main/java/entity/NguyenLieuEntity.java
@@ -1,7 +1,5 @@
 package entity;
 
-import java.math.BigDecimal;
-
 /**
  * Entity class:  NguyenLieuEntity
  */
@@ -9,7 +7,7 @@ public record NguyenLieuEntity(
         int id,
         String tenNguyenLieu,
         Double khoiLuong,
-        BigDecimal donGia,
+        Double donGia,
         int nhomThucPhamId
         ) {
 }
